Move compare.go into sub-package
